[#430] [#424] Ensure newly introduced n2s calls in the middle of parsing a trigger definition line do not affect parsing the remainder of the trigger definition by requiring input to trigger_parse() to never be in the stringpool
After integrating V63006 changes, we expected #424 to automatically be resolved since that seems to
have been addressed in V63006 even though no mention of that can be found in the V63006 release notes.
(see https://gitlab.com/YottaDB/DB/YDB/issues/424 for more details on the issue).

But we noticed the triggers/gtm7522a subtest still failing after the integration. This failure
happened only with a build of the YottaDB mainline. Never with GT.M V63006.

Turns out this is because #430 got merged into the YottaDB mainline and that introduced the
possibility of a n2s() call happening while in the middle of parsing the subscripts in a trigger
definition. This meant parsing the remainder of the trigger definition would get confused once
a stp_gcol() moves the input trigger definition (which lies in the stringpool) around.

To fix this issue, this commit now expects the passed in trigger definition string to trigger_parse()
to never be in the stringpool. This ensures it can safely do calls to n2s(). And to ensure that,
trigger_update() has a check of whether the trigger definition is in the stringpool. If it finds it
to be so, it takes a copy of the trigger definition into malloced memory and then passes this to
trigger_parse().