[#420] Ensure YDB signal handler is driven if SIGSEGV/SIGILL/SIGFPE/SIGBUS happens in SimpleThreadAPI process in thread holding YDB engine mutex lock while inside TP
This is an issue exposed by the newly developed simplethreadapi/random_walk subtest.
It encountered a SIG-11 due to an unrelated code issue but the SIG-11 did not generate a syslog
message (e.g. %YDB-F-KILLBYSIGSINFO1).

This was because of prior #420 changes. If we are in TP and get a SIG-11, the macro
FORWARD_SIG_TO_MAIN_THREAD_IF_NEEDED unconditionally assumed the signal needs to be forwarded
to another thread since it cannot safely determine the YDB engine mutex lock owning thread id.
While this is okay in case of SIGALRM (which always goes to the MAIN worker thread and hence
needs to be guaranteed forwarded to some other thread), it is not correct in case a thread
owning the YottaDB engine thread lock gets a SIG-11. In that case, it needs to continue signal
handling even if it is in TP. The macro is now fixed to handle this. After the changes,
the thread that got the SIG-11 executes generic_signal_handler() and dumps a core as expected
along with the %YDB-F-KILLBYSIGSINFO1 message in the syslog. Previously we saw a core dump only
but that was because we TIMED OUT (after a minute) in the DO_FORK_N_CORE_IN_THIS_THREAD_IF_NEEDED
macro while waiting for some other thread to signal us it is okay to dump core (not the correct way
to dump core). Now the core file is dumped almost immediately.