[DEBUG_ONLY] Fix possibility of SIG-11 in ydb_tp_s()/ydb_tp_st() if local variables are deleted inside the TP transaction
The simplethreadapi/random_walk subtest failed with a SIG-11 some times.
This happened only with dbg builds. Not with pro.

Below is the C-stack.

(gdb) where
 #0  __pthread_kill (threadid=<optimized out>, signo=3) at ../sysdeps/unix/sysv/linux/pthread_kill.c:57
 #1  0x00007f582c10815b in gtm_dump_core () at sr_unix/gtm_dump_core.c:72
 #2  0x00007f582c10967e in gtm_fork_n_core () at sr_unix/gtm_fork_n_core.c:163
 #3  0x00007f582c0fdca8 in generic_signal_handler (sig=11, info=0x7f582ce6e3c8 <stapi_signal_handler_oscontext+4296>,
     context=0x7f582ce6e448 <stapi_signal_handler_oscontext+4424>)
     at sr_unix/generic_signal_handler.c:348
 #4  <signal handler called>
 #5  0x00007f582c463597 in set_active_lv (newlv=0x55a6aff51490, do_assert=0, type=22)
     at sr_port/op_putindx.c:70
 #6  0x00007f582c52012a in tp_unwind (newlevel=0, invocation_type=COMMIT_INVOCATION, tprestart_rc=0x0)
     at sr_port/tp_unwind.c:139
 #7  0x00007f582c4713c8 in op_tcommit () at sr_port/op_tcommit.c:601
 #8  0x00007f582c2835f4 in ydb_tp_s_common (lydbrtn=LYDB_RTN_TP, tpfn=0x55a6adebcaf3 <gvnset>, tpfnparm=0x0, transid=0x0,
     namecount=0, varnames=0x0) at sr_unix/ydb_tp_s_common.c:281
 #9  0x00007f582c280398 in ydb_tp_s (tpfn=0x55a6adebcaf3 <gvnset>, tpfnparm=0x0, transid=0x0, namecount=0, varnames=0x0)
     at sr_unix/ydb_tp_s.c:44
 #10 0x000055a6adebca81 in main () at tp1.c:60

The set_active_lv() function is a dbg-only function. It was verifying that "newlv" was a valid lv.
It could not be in case of SimpleAPI use with TP since the most recent active local variable at the time
the TP started (ydb_tp_s()/ydb_tp_st() call) could have been killed inside the TP callback function.

In pro, set_active_lv() is not invoked and so we do not see the SIG-11. But it is still safer to ensure
active_lv is set back to only a valid lv at TP unwind time. So we follow the model that existed previously
in op_tstart.c and tp_unwind.c. active_lv used to be maintained only for implicit TP. Now that check is
further tightened to say if the TP was not started using SimpleAPI (all simpleapi TP is also implicit TP)
then do not maintain active_lv.

A simple test case to reproduce this issue is to do a ydb_set_s() of a local variable, then start a TP
transaction using ydb_tp_s() and inside the TP callback function do a ydb_delete_s() of that local
variable. In dbg, one will see the SIG-11 as above.