[V63006] ZCOMPILE "bar2.edit" should not compile bar2.edit.m but instead issue NOTMNAME error just like ZCOMPILE "bar2.edit.m" does
In GT.M V6.3-006, a NOTMNAME error is issued if a ZCOMPILE is attempted on a file name whose base name
is not a valid M routine name or if the file extension is not .m.

For example,

> touch bar2.edit.m
> mumps -run %XCMD 'ZCompile "bar2.edit.m"'
%YDB-E-NOTMNAME, bar2.edit is not a valid M name
%YDB-E-ZLNOOBJECT, No object module was produced

But if one tries the same without the .m extension, one does not get the same error.
Below is the output with V6.3-006.

> touch bar2.edit.m
> mumps -run %XCMD 'ZCompile "bar2.edit"'
> ls -1 bar2.o
bar2.o

A 0 sized object file bar2.o does get created. This is not correct if one interprets the following
release note at http://tinco.pair.com/bhaskar/gtm/doc/articles/GTM_V6.3-006_Release_Notes.html#GTM-8178

  In addition, the MUMPS and ZCOMPILE commands default file specifications without a .m file extension
  to have one and they only compile files with a .m extension

The code is now fixed to adhere to this release note. So a .m extension does get added to the
specified file name bar2.edit so bar2.edit.m is the actual file name that ZCOMPILE now sees. And
with this change, ZCOMPILE does issue a NOTMNAME error as below.

> mumps -run %XCMD 'ZCompile "bar2.edit"'
%YDB-E-NOTMNAME, bar2.edit is not a valid M name
%YDB-E-ZLNOOBJECT, No object module was produced

No separate issue created since this fix makes the code behave just like what the documentation says.