[#420] Add missing REVERT before returning in case of an error after an ESTABLISH_NORET (or else condition handler stack will keep growing indefinitely)
The simpleapi/random_walk subtest failed with the below fatal error.

%YDB-F-GTMASSERT2, YottaDB r998 Linux x86_64 - Assert failed sr_unix/condstk_expand.c line 52 for expression (new_size <= CONDSTK_MAX_STACK)

This turns out to be due to the use of ydb_exit() inside of the TP callback function in SimpleAPI mode.
That issues a INVYDBEXIT error which resulted in growth of the condition handler stack due to a missing
REVERT. Since the random_walk subtest does a lot of these operations in a random fashion, it eventually
expands the condition handler stack to a size more than the maximum design value of 512 resulting in
the fatal error failure.

This is now fixed by doing the REVERT not just in ydb_exit() but in a few other similar functions as well.