[DEBUG_ONLY] Remove incorrect (and unnecessary) assert in sr_unix/ydb_incr_s.c introduced in [#59]
There was an assert in ydb_incr_s.c which expected that the result of an op_add() will never
have the MV_STR bit set. But it is possible if the input node is an lvn and is not defined
and the increment is a string literal mval (MV_STR bit is already set), the result of the
ydb_incr_s() would be a copy of the string literal mval with the MV_STR bit set. That would
fail the assert (the recently enhanced go/random_walk subtest failed this assert).

Note that SET_YDB_BUFF_T_FROM_MVAL invoked at the end of ydb_incr_s() does a MV_FORCE_STR of the result
if it does not already have the MV_STR bit set so the assert served no purpose and is now removed..