[#449] Invalidate gv_target->prev_key if leaf block got rescanned as part of a recomputation or else $order(gvn,-1) or $query(gvn,-1) could cause SIG-11 or return incorrect results
gv_target->prev_key is maintained as part of scanning a leaf level block in a reverse $order or $query
operation. This is a performance enhancement introduced in GT.M V6.2-002. Turns out it is not maintained
correctly in case a recomputation (and rescanning) of the leaf level block occurs inside the commit logic
as part of sr_port/tp_tend.c or sr_port/gvincr_recompute_upd_array.c. Those re-scans update
gv_target->hist.h[0].prev_rec.offset but leave gv_target->prev_key unchanged leaving the two of them
out of sync. This means that any future operation which relies on prev_key (it is relied upon by the
reverse $order and $query operations in case gv_target->clue.end is non-zero) could return incorrect
results.

In the worst case, it is possible gv_target->prev_key->end is 0 (e.g. if a reverse $order
operation landed on the left most node in the GVT) at which point gv_target->hist.h[0].prev_rec.offset
would be 0. If a recomputation moves gv_target->hist.h[0].prev_rec.offset to a non-zero offset,
then a later use of prev_key would try to do a memcpy of prev_key into gv_altkey with a key length
equal to prev_key->end - 1 and since prev_key->end is 0, we would be trying to copy a memory section
of length -1 which when treated as an unsigned integer would be a huge positive integer causing
a massive overwrite of process memory and eventually resulting in a SIG-11.

This issue is now fixed by invalidating the prev_key in case a recomputation occurs in either
of sr_port/tp_tend.c or sr_port/gvincr_recompute_upd_array.c.

In addition, various other files have been changed to assert the validity of prev_key in debug
builds.