Fix rare possibility of SIG-11 while issuing DBCCERR, SYSCALL and FILEOPENFAIL errors
On a slow ARMV6L box, we saw a test failure in the manually_start/4g_journal subtest
where the update process reader helper got a SIG-11. This was because it got a
DBCCERR error because the timeout of 200 seconds was not enough to get the queue interlock
on the slow system but while issuing the DBCCERR error, it had passed the arguments
in the wrong order (LIT_AND_LEN usage instead of LEN_AND_LIT) and that caused the
SIG-11. The SIG-11 is now fixed by using LEN_AND_LIT.

Below is the C-stack of the process.

 #0  __pthread_kill (threadid=<optimized out>, signo=3) at ../sysdeps/unix/sysv/linux/pthread_kill.c:57
 #1  0xb5e9e384 in gtm_dump_core () at sr_unix/gtm_dump_core.c:72
 #2  0xb5e9fae0 in gtm_fork_n_core () at sr_unix/gtm_fork_n_core.c:163
 #3  0xb5e911c4 in generic_signal_handler (sig=11, info=0xb6bedad8 <stapi_signal_handler_oscontext+3528>, context=0xb6bedb58 <stapi_signal_handler_oscontext+3656>) at sr_unix/generic_signal_handler.c:377
 #4  <signal handler called>
 #5  0xb5ea6760 in gtm_wcswidth (ptr=0xa <error: Cannot access memory at address 0xa>, len=-1230382516, strict=0, nonprintwidth=1) at sr_unix/gtm_utf8.c:147
 #6  0xb5fec564 in util_format (message=0xbe9a961e "", fao=..., buff=0xedec30 "%YDB-F-SIGMAPERR, Signal was caused by an address not mapped to an object", size=2046, faocnt=0) at sr_unix/util_output.c:365
 #7  0xb5fef290 in util_out_print_vaparm ( message=0xbe9a95cc "%YDB-E-DBCCERR, Interlock instruction failure in critical mechanism for region !AD", flush=0, var=..., faocnt=2) at sr_unix/util_output.c:798
 #8  0xb5ea3a64 in gtm_putmsg_list (csa=0x0, arg_count=7, var=...) at sr_unix/gtm_putmsg_list.c:119
 #9  0xb5ea2208 in gtm_putmsg_csa (csa=0x0, argcnt=9) at sr_unix/gtm_putmsg.c:71
 #10 0xb63cb02c in updproc_preread () at sr_port/updhelper_reader.c:227
 #11 0xb63cac44 in updhelper_reader () at sr_port/updhelper_reader.c:139
 #12 0xb5e280d4 in mupip_main (argc=4, argv=0xbe9ac5a4, envp=0xbe9ac5b8) at sr_unix/mupip_main.c:124
 #13 0x000114ac in dlopen_libyottadb (argc=4, argv=0xbe9ac5a4, envp=0xbe9ac5b8, main_func=0x115f8 "mupip_main") at sr_unix/dlopen_libyottadb.c:148
 #14 0x00010b68 in main (argc=4, argv=0xbe9ac5a4, envp=0xbe9ac5b8) at sr_unix/mupip.c:19

While fixing this, a couple of similar issues were found in other places of the code and they are
also fixed.

Since all of these are rare error scenarios, a user-visible issue is not created for this.