[#452] Ensure the mval <arg_mval> is initialized in ydb_cij() and ydb_ci_exec() in all cases
Previously it was only initialized (using MV_INIT) if the call-in returned a value.
But even in the no-return case, arg_mval was being used in case a passed in parameter type
is "ydb_string_t *". In this case, not doing the MV_INIT of arg_mval could cause an
uninitialized mval to be later passed to the YottaDB runtime invocations later inside
the M program. If any of those invocations use the fnpc_indx field of the mval
(e.g. op_fnzpopulation()), they might behave non-deterministically. Thanks to a valgrind
run on the Octo executable for identifying this longstanding uninitialized variable case.