[#454] If ydb_routines/gtmroutines env var is not defined or set to "", set $zroutines to "$ydb_dist/libyottadbutil.so" if it exists and "$ydb_dist" otherwise
Note that previously the default value of $zroutines in this case was ".".  That is being changed
now. Although this is a backwards incompatible change with respect to GT.M, this is not considered
an issue because we expect most users to already set ydb_routines/gtmroutines env var to the value
they want as part of their YottaDB/GT.M startup scripts.