Fix possible buffer overflow with MUPIP BACKUP with destination directory full path lengths close to 255 bytes
The v63003/gtm4212 subtest failed when YottaDB was built with clang (normally it is built with gcc)
with the below output.

gtm4212.log
 # Backing up DEFAULT Region to path length 228 (length of temp file is 24, so total path is 252)
 /bin/mv: cannot stat 'aa...aa/DEFAULT_51648__0.tmp/mumps.'$'\001\003\001': No such file or directory

"aa...aa" above represents a long directory name where 'a' repeats 228 times.

The file name extension should have been .dat (i.e. mumps.dat) but instead shows up as garbage ('$'\001\003\001).
This is because we have a buffer overflow issue that gcc was forgiving about but clang is not and helped
identify.

We initially create the temporary file 'aa...aa/DEFAULT_51648__0.tmp/mumps.dat' using a buffer "cmdarray"
whose size is 1024. So that fit in the above name fine. But later we try to open this file name using
the macro OPENFILE by passing a buffer "tempfilename" which is just 256 bytes long. Just before the OPENFILE
we copy the mumps.dat portion into tempfilename[] and that causes it to become 258 bytes long causing a
3-byte buffer overflow. The OPENFILE succeeds even with the buffer overflow but later we use tempfilename[]
by copying that into command[cmdlen] in a memcpy done as part of a "mv" command. And while doing that
for some reason, the overflow bytes "dat" become garbage. Not sure exactly what causes this but that is
not important.

The buffer overflow is now fixed by ensuring "cmdarray" is the same size as "tempfilename". This ensures
whatever file we were successfully able to create using "cp" and "cmdarray" can be opened without a buffer
overflow using "tempfilename". And this ensures no later issues when we use "tempfilename" to do the
memcpy for the "mv" command.