[#431] LVUNDEF error from ydb_get_s()/ydb_get_st() displays variable name with subscripts (if any) in the buffer returned by ydb_zstatus()
Some background first. LVUNDEF error inside a ydb_get_s()/ydb_get_st() call could previously be
issued in any one of the below places.
	(a) libyottadb_int.h as part of the FIND_BASE_VAR_NOUPD macro.
	(b) op_getindx() called from ydb_get_s()
	(c) ydb_get_s() if LV_IS_VAL_DEFINED(src_lv) is FALSE.

(a) and (c) were issuing the LVUNDEF error without even the variable name and/or subscripts.
(b) issued the LVUNDEF error with subscripts but without the variable name. This is because it
	expects the variable name to be stored in the local variable name hashtable which is
	not done for the ydb_get_s() SimpleAPI call (is done only for M access).

This is now fixed by removing the LVUNDEF from (a). This was okay since the only caller of the
FIND_BASE_VAR_NOUPD macro that passed ERR_LVUNDEF_OK_TRUE was ydb_get_s() so that parameter itself
is now removed and all callers of the macro correspondingly adjusted to call with one less parameter.

As for (b), we do not want op_getindx() to issue LVUNDEF error in case it is called from ydb_get_s().
This is because it has no clue about the variable name. Therefore op_getindx() is changed to check
if the caller is ydb_get_s() (which it can using TREF(libyottadb_active_rtn)) and if so return
&literal_null in the lv in the LVUNDEF case.  The caller ydb_get_s() now checks this special return
and if so issues the LVUNDEF error right there (this replaces the LVUNDEF error previously issued
as part of (c)).

ydb_get_s() is also enhanced to issue the LVUNDEF error by printing string subscripts in $zwrite
format (i.e. enclosed inside double-quotes, use $ZCHAR/$CHAR syntax in case of binary characters etc.)
and print numeric subscripts as is (i.e. no double-quotes etc.). Note that this is an enhancement
compared to the LVUNDEF error issued by M code (in format_key_lv_val.c) where all string subscripts
used to be displayed without being surrounded by double quotes. In addition, ydb_get_s() takes care
to truncate the display in case the subscripts take up more than 512 bytes of space in the output
buffer. Previously we could get garbage output.

The LVUNDEF error printing logic for M code in format_key_lv_val.c is left untouched as it might
be a backward incompatible enhancement if we were to change that too.