Work around weird Clang compiler issue that causes SIG-11/GTMASSERT2 failures in pro builds
The v62000/gtm8108 and v55000/gtm7020 subtests failed when run with the pro builds if the
build was done using the Clang/LLVM compiler (recently supported in YottaDB).

The failure symptoms were one of the below

> %YDB-F-GTMASSERT2, YottaDB r999 Linux x86_64 - Assert failed sr_port/stp_gcol_src.h line 834 for expression (FALSE && mvs->mv_st_type)
> %YDB-F-KILLBYSIGSINFO1, YottaDB process 27817 has been killed by a signal 11 at address 0x00007FB58747CD1A (vaddr 0x00000000029929AB)

After some analysis, it was found that the variable "val" was getting set to an off-by-8 value
(i.e. 8 bytes more than what it should be) in gvcst_spr_queryget.c. Not at all clear why or how.
Clang does not seem to offer a source listing option either. So cannot confirm this but my suspicion
is that Clang is generating incorrect code in this case. When I ensure the local variable "val"
is set to a dummy global variable, the problem disappears. Since this takes care of both the
above subtest failures and it entails one extra pointer global variable ("dummy_ptr"), we consider
this good enough of a workaround.