Enhance ydbinstall to handle alphanumeric YottaDB release names (possible in development environment)
This is not an issue with production releases of YottaDB which have the form rN.NN where N is
a number. In development though, we could build say a release r9.98A. In that case, the
logic in ydbinstall which checks if pkg-config file install needs to happen does a comparison
of whether a higher release is already installed using a numeric comparison operator (-lt).
This fails with alphanumeric release names so the logic is now reworked to do string comparison
instead.

Also, grep -P is an experimental feature and is better avoided so an equivalent expression is
constructed without using -P.