[#451] Increase ydb_stm_thread() sleep time from 1 microsecond to 1 second to reduce CPU usage of this mostly idle thread
Previously the sleep time was kept at 1 microsecond to ensure a SIGALRM signal is handled in a timely
fashion (i.e. as soon as it is received). But in looking at it again, that should not be a big concern.
The SIGALRM signal is going to interrupt the sleep. Therefore as long as we pass a FALSE parameter to the
SLEEP_USEC macro, the macro will return right away when the SIGALRM is received letting us handle/forward
the signal to the appropriate thread in a timely fashion with no loss of time compared to the 1-micro-second
sleep approach.

Since 1-micro-second sleep loop eats more CPU time then a 1 second sleep loop, the loop sleep time
is now bumped to 1 second.

This simplified a lot of the logic that previously existed mainly because the sleep loop was in microseconds
and we wanted to avoid waking up the YDB engine multi-thread lock holding thread too frequently (i.e. less
than 1 second apart). Since we now wait for 1 second, we can safely wake up the holding thread
without checking how far back in time we tried to wake up the same thread. This meant a lot of the prior
code is no longer necessary and simplified the code noticeably.

Also took this opportunity to modify the comments before the function definition to better describe the
purpose of this function after the #420 changes.