[V63007] Restore changes to GTM_FWRITE()/gtm_fwrite() that were lost in V63007 conflict resolution
A prior commit 99b3808e0fd35e017965c3c5cdd7b52948ca072e had made changes to gtm_fwrite()
to ensure the return value RC in the GTM_FWRITE macro contained the "errno" in case of an error.
This change was lost when V63007 changes were integrated. It is now restored.

In addition, an assert was introduced in V63007 in gtm_fprintf() (in sr_unix/gtm_stdio.c) that
retval is never -1. But it is possible for this scenario if the principal device goes away
in the middle like is the case in the r124/ydb95 subtest (where a MUPIP EXTRACT output is piped
to a head -2 which closes the principal device as soon as 2 lines are seen and so the MUPIP EXTRACT
would get an error from gtm_fprintf() which is expected. Therefore, the newly added assert in V63007
is now removed.