[DEBUG-ONLY] M OPEN command with timeout should cancel timer if Ctrl-C aborts it
We noticed this issue in a DBG build of YottaDB. The following test attempts an OPEN command
with 60 second timeout but aborts it using a Ctrl-C before the timeout expires and then attempts
the same OPEN command again. An assert fails indicating an internal timer that had been started
for the first OPEN command was not canceled before returning to the YDB> prompt.

YDB>OPEN "unicodefifoM.pipe":(FIFO:WRITE:RECORDSIZE=-1:CHSET="M"):60
^C%YDB-I-CTRLC, CTRL_C encountered

YDB>OPEN "unicodefifoM.pipe":(FIFO:WRITE:RECORDSIZE=-1:CHSET="M"):60
%YDB-F-ASSERT, Assert failed in sr_unix/gt_timers.c line 949 for expression (NULL == find_timer(tid, &tpp))

We did not notice any issues in the PRO build. This is because in start_timer_int() in gt_timers.c
there is code that invokes a remove_timer() if not a DBG build.