[CLANG] Work around weird Clang compiler issue that results in incorrect TPFAIL message due to garbage function return values
The triggers/trigdefbad subtest failed in pro builds (with optimization turned on) when built using
Clang (as opposed to GCC).

The failure diff was the following.

97c97
< %YDB-E-TPFAIL, Transaction COMMIT failed.  Failure code: ZZZZ.
---
> %YDB-E-TPFAIL, Transaction COMMIT failed.  Failure code: ZZZ=.

The failure code in the final retry was '=' instead of 'Z'. After some analysis it was found
that the function gvcst_search_blk() was returning a garbage value of 0x7830203d (2016419901 decimal)
from the if block at lines 269-272 below when t_tries == 3. Whether t_tries < 3 or == 3, the
return value in all cases is expected to be cdb_sc_blklenerr (i.e. 'Z').

sr_port/gvcst_blk_search.h
--------------------------
    266    if (pRec >= pTop)
    267    {       /* Terminated at end of block */
    268            if (pRec > pTop)        /* If record goes off the end, then block must be bad */
    269            {
    270                    INVOKE_GVCST_SEARCH_FAIL_IF_NEEDED(pStat);
    271                    return cdb_sc_blklenerr;
    272            }

When t_tries < 3, the macro INVOKE_GVCST_SEARCH_FAIL_IF_NEEDED() turns out to be a no-op as it has
a check for if t_tries == 3 and only in that case it invokes the function gvcst_search_fail().

Below is how the macro is currently defined.

 #define INVOKE_GVCST_SEARCH_FAIL_IF_NEEDED(pStat)  if (CDB_STAGNATE <= t_tries) gvcst_search_fail(pStat);

For reasons not yet clear, the clang compiler with optimization flags turned on ends up generating
incorrect code that if t_tries == CDB_STAGNATE (i.e. 3) invokes gvcst_search_fail(pStat) in line 270
above correctly but does not go to line 271 and instead returns some other unrelated value. This 4-byte
value has the last byte value as 0x3d which is '=' and that is what finally shows up in the TPFAIL message.

To workaround the issue, the macro definition is changed as follows. The entire macro is now enclosed
within braces. That somehow causes the compiler to generate the right code.

 #define        INVOKE_GVCST_SEARCH_FAIL_IF_NEEDED(pStat)       \
 MBSTART {							\
        if (CDB_STAGNATE <= t_tries)				\
                gvcst_search_fail(pStat);			\
 } MBEND