[#382] [V63006] Fix LKE CLNUP to cleanup all orphaned locks
GT.M V6.3-006 (which is when the LKE CLNUP functionality was introduced) could return
after cleaning only some of the orphaned locks. We noticed this while testing #382.
This is now fixed.

There was code that went through the right-sibling of a mlk_shrsub structure in the
fill_pid_table() function. The variable "d2" was being maintained every iteration of
the do/while loop by going 2 nodes to the right. Not sure why that was done.
Since d2 == d was the loop terminating condition, and since d2 traversed the rsib linked
list a lot faster than d, we terminated the loop when d2 came around to d but before d had
gone through all the nodes in the rsib linked list. d2 is now gotten rid of and that
fixes the issue.