Handle a restartable situation (empty index block) in gvcst_search_blk()
We had a test failure in the resil/resil and reorg/on_ntp_njnl_reorg subtests.
Both with the following symptoms.

> %YDB-F-ASSERT, Assert failed in sr_port/gvcst_blk_search.h line 347 for expression (pStat->curr_rec.offset >= SIZEOF(blk_hdr))

This is a case where what was an index block is empty whereas we expect at least a star key in it.
In this case, we were about to return from gvcst_search_blk() with pStat->curr_rec.offset == 0.
The assert prevented that from happening as it expects us to return a curr_rec with offset at
least SIZEOF(blk_hdr).

Clearly, this is a restartable situation (the index block is no longer an index block) and so
it is best that we restart. That is what is done now.

From the code, it is not clear how callers will handle a return from gvcst_search_blk() where
curr_rec.offset is 0. But given both failures till now have been only in debug builds, it is hoped
there is no issue in production builds. Hence no issue created for this commit.

One notable thing is that both failures happened in debug builds of YottaDB using CLang as the
C compiler (instead of GCC). Not sure if that is a coincidence or has a relationship. Wanted to
record it just in case it helps at a later point in time.