[#463] Fix possibility of SIG-11 if SimpleThreadAPI and using multiple threads inside ydb_tp_st()
See https://gitlab.com/YottaDB/DB/YDB/issues/463 for background on the issue.

The address of the zeroes4byte variable (which was in the C-stack) was being copied over to the
update array as part of a BLK_SEG operation. Later this update array could be used by another
thread while inside the same TP transaction to build a newer version of this block. At that point
if the original thread where the BLK_SEG happened no longer exists, it is possible the pointer
stored in the update_array as part of the BLK_SEG is no longer accessible and it can cause a SIG-11
in gvcst_blk_build(). This was seen in the simplethreadapi/randomWalk subtest.

The fix is to not use a pointer to the C-stack but instead copy it from the C-stack to the update-array
(using a BLK_ADDR operation) while in the first thread. This way a later thread will find all that
it needs in the update-array itself. No pointers to other thread memory.