[#429] Fix issue with recording live semaphores (in "sem" local variable) from ipcs -s output
The code was attempting to filter out lines from the ipcs -s output that contained
a key of 0x00000000 (IPC_PRIVATE i.e. private semaphore). To do this it was incorrectly
looking for 0x... in the 2nd column (semid). Instead it should have been looking at the
first column (semkey).

The consequence of this issue was that ydb_env_set would incorrectly try to run a rollback/recover
on a live database.

This is now fixed by looking at the 1st column for the semkey.