[DEBUG-ONLY] Fix MUPIP REORG -TRUNCATE to restart the transaction correctly in a rare case
We had the resil/resil subtest fail with the following diff.

28a29,75
> resil_4_15/resil/reorg4.out
> %YDB-F-ASSERT, Assert failed in sr_port/t_retry.c line 161 for expression ((cdb_sc_normal != failure) || mu_reorg_process)

The C-stack at the time of the core was the following.

 () where
 #0  pthread_kill () from /usr/lib64/libpthread.so.0
 #1  gtm_dump_core () at sr_unix/gtm_dump_core.c:74
 #2  gtm_fork_n_core () at sr_unix/gtm_fork_n_core.c:163
 #3  ch_cond_core () at sr_unix/ch_cond_core.c:80
 #4  rts_error_va () at sr_unix/rts_error.c:192
 #5  rts_error_csa () at sr_unix/rts_error.c:99
 #6  t_retry () at sr_port/t_retry.c:161
 #7  mu_truncate () at sr_unix/mu_truncate.c:215
 #8  mupip_reorg () at sr_port/mupip_reorg.c:401
 #9  mupip_main () at sr_unix/mupip_main.c:123
 #10 dlopen_libyottadb () at sr_unix/dlopen_libyottadb.c:148
 #11 main () at sr_unix/mupip.c:19

 (gdb) f 7
 #7  0x00007f616a13921d in mu_truncate (truncate_percent=0) at sr_unix/mu_truncate.c:215
 215      t_retry((enum cdb_sc)rdfail_detail);

 211     bmphist.buffaddr = t_qread(bmphist.blk_num, &bmphist.cycle, &bmphist.cr);
 212     lmap_blk_hdr = (blk_hdr_ptr_t)bmphist.buffaddr;
 213     if (!(bmphist.buffaddr) || (BM_SIZE(BLKS_PER_LMAP) != lmap_blk_hdr->bsiz))
 214     { /* Could not read the block successfully. Retry. */
 215             t_retry((enum cdb_sc)rdfail_detail);
 216             continue;
 217     }

In the case where bmphist.buffaddr is NULL, t_qread() would set rdfail_detail to a status code that
is not cdb_sc_normal. But in case bmphist.buffaddr is non-NULL, rdfail_detail would be untouched and
so it would be cdb_sc_normal. Therefore, if lmap_blk_hdr->bsiz does not match the size of a bitmap
block, we would be invoking t_retry() with a status code of cdb_sc_normal which t_retry is not
designed to handle. It fails an assert that the only known exception to this rule is if mupip reorg
calls it. In this case, it is not the reorg phase of mupip reorg but the later truncate phase of
mupip reorg that calls it. And so the assert is still valid. The issue is with mu_truncate.c and
it needs to return an abnormal status code in the latter case. This is now fixed to return cdb_sc_lostcr
and an assert added that this can happen only in BG access method (buffer reuse).

In a pro build, I cannot see anything that will break inside t_retry.c when cdb_sc_normal is passed.
Seems like it will still retry the transaction which is the right thing to do. So this is almost likely
a non-user-visible issue no gitlab issue is created.