[#429] Fix errors found in ydb_env_set when parsing /proc/self/environ with blank lines and when shared object does not exist
GetCurrEnv doesn't account for getting a blank line from
`/proc/self/environ` and stops processing environment variables
afer a blank line is encountered. This is fixed by reading until
an EOF is encountered. Blank lines are then skipped when converting
the environment variables into an array.

Also, fix a case where if the shared object doesn't exist a blank
space would be added to the end of ydb_routines/gtmroutines.