[V63007] Fix rare v43001/zjob subtest failure (due to GTM-9058 changes in GT.M V6.3-007)
The v43001/zjob subtest failed with the following diff on a 32-bit ARMV6L box (with 1-CPU).

```
18c18
< %YDB-E-JOBFAIL, JOB command failure,%YDB-I-TEXT, Job error in routine specification. Label and offset not found in created process
---
> %YDB-E-JOBFAIL, JOB command failure,%YDB-I-TEXT, Job error in routine specification. Label and offset not found in created process,%SYSTEM-E-ENO4, Interrupted system call
```

After some analysis, this turns out to be a code issue that can cause an extraneous SYSTEM-E-... error
message dispalyed as part of the JOBFAIL error in rare cases.

I suspect what happened is related to GTM-9058 code changes in GT.M V6.3-007
(http://tinco.pair.com/bhaskar/gtm/doc/articles/GTM_V6.3-007_Release_Notes.html#GTM-9058 has the
corresponding release note). Lines 561-570 below was introduced as part of that change.

```
sr_unix/ojstartchild.c
----------------------
   561   /* Read status to catch any basic errors */
   562   DOREADRC(setup_fds[0], &joberr, SIZEOF(joberr), rc);
   563   if (rc < 0)
   564           joberr = joberr_rtn;    /* Assume routine error if there is a problem getting the report */
   565   if (joberr_ok != joberr)
   566   {
   567           job_errno = errno;
   568           DOWRITERC(pipe_fds[1], &job_errno, SIZEOF(job_errno), pipe_status);
   569           UNDERSCORE_EXIT(joberr);
   570   }
```

The DOREADRC macro (which is executed in the middle child process) read "joberr" from the grandchild
process as the value "joberr_rtn" because the grandchild got a JOBLABOFF error (the test intentionally
creates this error) and returned with "rc" greater than 0. But as part of the DOREADRC macro
(which executes a read() system call), it got an EINTR error return which got handled by a retry
of the read() in the loop all within the macro until there was no error return from read(). But the
value of "errno" when DOREADRC returned was set to EINTR because of this and so at line 567 above,
job_errno got incorrectly set to EINTR. This resulted in the final error message issued to have the
"%SYSTEM-E-ENO4, Interrupted system call" suffix in the JOBFAIL error message.

The fix is to set job_errno to errno only if "rc" is < 0.

In addition, an assert has been added to op_job.c to verify that if ever the exit status of the middle
child is joberr_rtn, then job_errno is always -1 (i.e. never corresponds to a system error like EINTR
in the test failure case).

Since V6.3-007 was integrated after r1.24 and was never part of an official YottaDB release, there is
no need to create a user-visible issue for this.